SwingSearchBar: encapsulate some functions
There are a couple of operations -- showing and hiding the panel,
as well as losing the focus -- where it might be nice for
downstream code to be able to override or tweak the behavior.

This commit returns the behavior to using a child dialog,
since it makes fewer assumptions about the containing parent.

However, with the protected showPanel and hidePanel methods,
it is now possible to customize the behavior such that the
panel can be embedded differently as desired.