Fix infinite loop when search starts with a space
When the search string begins with a space, one of the terms becomes the
empty string. When you call s.indexOf("", s.length() + 1) -- i.e., out
of bounds of the string -- the indexOf routine returns s.length()
instead of -1. So we must check that index is in bounds of the string on
both sides.

This commit is dedicated to Niko Ehrenfeuchter.