Add a javadoc button for class search results
To achieve this, there is now a JavadocService that knows how to
provide a remote URL string for a given fully qualified class name.
It works by scraping the javadoc.scijava.org index page, and then
linked projects' overview-summary pages, to obtain package lists.
Since these accesses takes some time, it is done proactively on
dedicated threads when the JavadocService is first initialized.