Enlarge search result icon size to 32x32
The assumption was that search result icons would always be 16x16,
which is the long-time favicon size. But with HiDPI displays in
common use now, it's better to have more detail. Ultimately, this number
should stop being hardcoded and instead be derived from the UI scale or
some such, but for the moment, just upping it to 32x32 is still better.